refactor(core): resolve circular dependency
Since in the previous commit db80d15, `BasicEditor` uses
`RichTextEditor` to render draft editor. The dependencies
will be like:
- `RichTextEditor` imports `Annotation`, `InfoBox`
- `Annotation` and `InfoBox` import `BasicEditor`
- `BasicEditor` imports `RichTextEditor`

There is a circular dependency between `RichTextEditor` and `BasicEditor`.

In order to resolve the circular dependency, this patch removes
`BasicEditor`. Instead, it uses render props to render `RichTextEditor`
in `Annotation` and `InfoBox`.