feat(seinsights): update SpecialfeatureList.ts. Record the user input order of relationship `specialfeatures` field
For relationship fields, KeystoneJS won't take input order into account.
Therefore, if we want to remember the input order, we need to do it by
ourselves.

This commit adds `manualOrderOfSpecialFeatures` field, which is used to
record the input order of `specialfeatures` field.

We monitor the input order in `resolveInput` hook.
If the order changes, we adjust `manualOrderOfSpecialFeatures` field
correspondingly.